fix(ffi): read size_t out-params at their native width — ILP32 armv7
Every size_t out-param was read as uint64. On the LP64 platforms
they coincide; on arm-linux (ILP32) the read ran out of bounds —
the 0.6.18.1 armv7 smoke died in ValueML.load_all_marshal
(FFI::AbstractMemory#read_uint64 offset=0 size=8 out of bounds)
after the build itself succeeded. The ffi gem (1.17.4) ships no
read_size_t, so Yeptris::FFI.read_c_size_t dispatches on the
measured type size; the three :uint64 allocations that only worked
by zeroed-little-endian luck align to :size_t as well.

+++ lib/yeptris/document.rb
@@ -164,7 +164,7 @@ class Yeptris::Document
 
   def serialize(canonical: false, best_width: 0, explicit_doc_start: false)
     ensure_alive!
-    len = ::FFI::MemoryPointer.new(:uint64)
+    len = ::FFI::MemoryPointer.new(:size_t)
     ptr =
       if canonical || best_width.positive? || explicit_doc_start
         opts = Yeptris::FFI::EmitOptions.new
@@ -180,13 +180,13 @@ class Yeptris::Document
   end
 
   def serialize_json_compact
-    len = ::FFI::MemoryPointer.new(:uint64)
+    len = ::FFI::MemoryPointer.new(:size_t)
     Yeptris::FFI::Owned.string(Yeptris::FFI.yeptris_serialize_json_ex(@c_ptr, len, 1), len)
   end
 
   def serialize_json
     ensure_alive!
-    len = ::FFI::MemoryPointer.new(:uint64)
+    len = ::FFI::MemoryPointer.new(:size_t)
     Yeptris::FFI::Owned.string(Yeptris::FFI.yeptris_serialize_json(@c_ptr, len), len)
   end
